No auto_inc of gorm:"primary_key" if foreignkey
If you are trying to create a table to store metadata about the
relationship of two entities, e.g. Product and Store and the quantity of
product available at that store. You may wish to use a composite foreign
key relationship.

If you also wish to use gorm to generate the SQL to create the schema,
but default gorm will add the auto_increment keyword to tables which are
both primary and integer. This is a undesirable for composite foreign
key relationships and generates invalid sql statements.

If you have a struct field which is a primary key and is marked as a
foreign key, disable the auto increment behavior while generating the
schema.