sqlite3: Never set primary key in-line
Never set the primary key flag as part of the column definition. When
using sqlite3, this prevents mixed int/other type composite primary
key (because the int column always wins and get the index for itself).

Rather, let's just set the primary key at the end, whether it's composite
or not. Note that the int key will automatically be AUTOINCREMENT if
its type is "integer". For that reason, always use that type, rather
than "bigint", as bigint maps to "integer" eventually anyway.