Treat partially zero compound foreign keys as an unset association
Currently when one value of a compound foreign key is null or zero,
preloading this association yields a panic ("reflect: call of
reflect.Value.Interface on zero Value") because we're attempting to
look up the identity values when any one of the columns is non-zero.

The fix here is to require each of the values to be non-zero or we
assume that the association is missing. This is not a very common
scenario to have in practice, but it's an edge case that's easy to fix
and resolves a use-case for me.